pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Educational project. Determines differences between two files"
authors = ["Timur <tim.vildanov@gmail.com>"]
readme = "README.md"
packages = [{include = "hexlet_code"}]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
flake8 = "^6.1.0"
path = "^16.7.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
